Forenkler utbetalingsperioder fra vedtak
